+++ src/test/kotlin/com/atlassian/performance/tools/infrastructure/browser/SshChromium.kt
@@ -1,7 +1,7 @@
 package com.atlassian.performance.tools.infrastructure.browser
 
-import com.atlassian.performance.tools.infrastructure.api.jvm.StaticBackoff
 import com.atlassian.performance.tools.jvmtasks.api.IdempotentAction
+import com.atlassian.performance.tools.jvmtasks.api.StaticBackoff
 import com.atlassian.performance.tools.ssh.api.DetachedProcess
 import com.atlassian.performance.tools.ssh.api.SshConnection
 import com.atlassian.performance.tools.virtualusers.api.browsers.Browser